quest recommendation helper methods
