+++ src/core/quest/getRecommendedSkills.ts
@@ -1,0 +1,31 @@
+import { extractTemplatesFromString } from '../../utils/wikimedia/extractTemplatesFromString'
+
+export const getRecommendedSkills = (recommended: string) => {
+	// Parse recommended skills from the recommended string
+	const recommendedSkills: Record<string, number> = {}
+
+	// Extract SCP templates from the recommended string
+	const templates = extractTemplatesFromString(recommended)
+
+	// Filter for SCP templates (skill clickpic templates) - exact match only
+	const skillTemplates = templates.filter(template => template.title === 'SCP')
+
+	// Process each SCP template to extract skill and level
+	for (const template of skillTemplates) {
+		const skillName = template.parameters['1']?.trim()
+		const levelStr = template.parameters['2']?.trim()
+
+		if (skillName && levelStr) {
+			const level = parseInt(levelStr, 10)
+			if (!isNaN(level)) {
+				// Store the recommended skill level
+				// If the same skill appears multiple times, keep the highest
+				if (!recommendedSkills[skillName] || recommendedSkills[skillName] < level) {
+					recommendedSkills[skillName] = level
+				}
+			}
+		}
+	}
+
+	return recommendedSkills
+}
